pyproject.toml
[tool.poetry]
name = "matches"
version = "0.1.0"
description = "Pytorch training toolkit for power users"
authors = ["Vsevolod Poletaev <poletaev.va@gmail.com>", "Nikolay Kozyrskiy <nik.kozyrskiy@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.0"
torch = ">=1.6.0"
pytorch-ignite = "^0.4.2"
tqdm = "^4.56.0"
coloredlogs = "^15.0"
tensorboardX = "^2.1"
wandb = "^0.13.5"
pandas = ">=1.4.4"
GitPython = {version = "^3.1.12", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
